skills/gemini-image-generator/scripts/requirements.txt
# Gemini Image Generator Dependencies
google-genai>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
